+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- Signature in url instead of in SAMLRequest body
+
 ### Fixed
 - Metadata rack response body is now an array
 - Typo on authn context comparison value
 - AllowCreate is not expected in NameIDPolicy element
+
+
 ## [0.9.0] - 2017-07-31
 ### Added
 - Rack middleware that handles spid login requests